ThriftnDrift/ThriftnDriftDataCollection/Scripts/requirements.txt
python-dotenv==1.0.0
googlemaps==4.10.0
firebase-admin==6.2.0
tqdm==4.66.1
python-google-places==1.4.2 